errorfixes

+++ src/main/java/sk/tuke/kpi/oop/game/Reactor.java
@@ -3,9 +3,7 @@ package sk.tuke.kpi.oop.game;
 import sk.tuke.kpi.gamelib.Scene;
 import sk.tuke.kpi.gamelib.framework.AbstractActor;
 import sk.tuke.kpi.gamelib.graphics.Animation;
-import sk.tuke.kpi.oop.game.action.PerpetualReactorHeating;
-import sk.tuke.kpi.oop.game.tools.FireExtinguisher;
-import sk.tuke.kpi.oop.game.tools.Hammer;
+import sk.tuke.kpi.oop.game.actions.PerpetualReactorHeating;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -41,7 +39,7 @@ public class Reactor extends AbstractActor implements Switchable, Repairable{
         return this.temperature;
     }
 
-    public float getDamage(){
+    public int getDamage(){
         return this.damage;
     }
